pyproject.toml
[tool.poetry]
name = "ddrr"
version = "0.1.0"
description = ""
authors = ["Deniz Dogan <denizdogan@users.noreply.github.com>"]

[tool.poetry.dependencies]
attrs = "^19.1"
Django = [
    {version = "^1.11", python = "~2.7"},
    {version = "^2.0", python = "^3.5"},
]
lxml = {version = "^4.3",extras = ["xml"]}
python = "^2.7.9 || ^3.5"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", allows-prereleases = true, python = "^3.6"}
faker = "^1.0"
faker_web = "^0.3.1"
flake8 = "^3.7"
flake8-black = {version="^0.0.4", python = "^3.6"}
flake8-bugbear = { version = "^19.3", python = "^3.5"}
flake8-isort = "^2.7"
isort = "^4.3"
poetry = "^0.12.15"
pre-commit = "^1.16"
pytest = "^4.5"
pytest-cov = "^2.7"
pytest-django = "^3.4"
pytest-mock = "^1.10"
pytest-runner = "^4.4"
tox = "^3.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
